add new features and fix the old ones

+++ app/Models/Client.php
@@ -20,12 +20,14 @@ class Client extends Model
 
     public function disconnect()
     {
-        $this->load('server');
-        if(is_null($this->server)) {
-            $this->load('serverUser.server');
+        if ($this->stop !== null) {
+            return false;
         }
+
+        $this->loadMissing('serverUser.server');
+
         $proto = app()->isLocal() ? "http" : "https";
-        $hostname = app()->isLocal() ? "stream:1985" : $this->server->hostname;
+        $hostname = app()->isLocal() ? "stream:1985" : $this->serverUser->server->hostname;
         Http::withBasicAuth(config('services.srs.username'),config('services.srs.password'))->delete($proto."://".$hostname.'/api/v1/clients/'.$this->client_id);
         return true;
     }
